[ALPS04783226] drcc: fix build break when CONFIG_KASAN=y
Fix build breaks when CONFIG_KASAN=y with CC=gcc

drivers/misc/mediatek/base/power/drcc_v1/mt6885/mtk_drcc.c:531:45:
error: format ‘%llx’ expects argument of type
‘long long unsigned int’,
but argument 3 has type ‘int’ [-Werror=format=]
rivers/misc/mediatek/base/power/drcc_v1/mt6885/mtk_drcc.c:1099:45:
error: format ‘%llx’ expects argument of type
‘long long unsigned int’,
but argument 3 has type ‘int’ [-Werror=format=]

MTK-Commit-Id: d7824ee030129e46dd0d9e2d4d1bf6140a949e0e

CR-Id: ALPS04783226
Feature: PTP-OD (Performance Thermal Power-Over Drive)

+++ drivers/misc/mediatek/base/power/drcc_v1/mt6885/mtk_drcc.c
@@ -529,7 +529,7 @@ static int drcc_trig_proc_show(struct seq_file *m, void *v)
 	}
 
 	seq_printf(m, "MCUSYS_RESERVED_REG1 : 0x%llx = 0x%x\n",
-		MCUSYS_RESERVED_REG1,
+		(unsigned long long)MCUSYS_RESERVED_REG1,
 		mt_secure_call_drcc(
 			MTK_SIP_KERNEL_DRCC_READ,
 			MCUSYS_RESERVED_REG1,
@@ -1097,7 +1097,7 @@ static int drcc_reg_dump_proc_show(struct seq_file *m, void *v)
 	}
 
 	seq_printf(m, "MCUSYS_RESERVED_REG1 : 0x%llx = 0x%x\n",
-		MCUSYS_RESERVED_REG1,
+		(unsigned long long)MCUSYS_RESERVED_REG1,
 		mt_secure_call_drcc(
 			MTK_SIP_KERNEL_DRCC_READ,
 			MCUSYS_RESERVED_REG1,
